Decomp types into kbcmf.go
